# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Freelancer Platform API
  description: |
    Comprehensive API for a global freelance marketplace with secure payments, 
    verified profiles, milestone-based delivery, and dispute resolution.
  version: 1.0.0
  contact:
    name: NivixPe Team
    email: api@freelancerplatform.com
servers:
  - url: https://api.freelancerplatform.com/v1
    description: Production
  - url: https://staging-api.freelancerplatform.com/v1
    description: Staging
  - url: http://localhost:3000/v1
    description: Local Development
...
security:
  - BearerAuth: []
paths:
...
  /disputes:
    post:
      tags: [Disputes]
      summary: File dispute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contractId, reason]
              properties:
                contractId:
                  type: string
                  format: uuid
                reason:
                  type: string
                evidence:
                  type: object
      responses:
        '201':
          description: Dispute filed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List all users (admin only)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /admin/users/{userId}/verify:
    post:
      tags: [Admin]
      summary: Verify user (admin only)
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User verified
  /admin/disputes/{disputeId}/resolve:
    post:
      tags: [Admin]
      summary: Resolve dispute (admin only)
      parameters:
        - name: disputeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution]
              properties:
                resolution:
                  type: string
                refundAmount:
                  type: number
      responses:
        '200':
          description: Dispute resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'
  /admin/analytics:
    get:
      tags: [Admin]
      summary: Get platform analytics (admin only)
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  gmv:
                    type: number
                  totalProjects:
                    type: integer
                  totalUsers:
                    type: integer
                  activeContracts:
                    type: integer

# ==================== COMPONENTS ====================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [FREELANCER, CLIENT, ADMIN]
        fullName:
          type: string
        avatarUrl:
          type: string
        profileData:
          type: object
        rating:
          type: number
        totalReviews:
          type: integer
        isVerified:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Dispute:
      type: object
      properties:
        id:
          type: string
          format: uuid
        contractId:
          type: string
          format: uuid
        filedBy:
          type: string
          format: uuid
        reason:
          type: string
        evidence:
          type: object
        status:
          type: string
          enum: [OPEN, UNDER_REVIEW, RESOLVED, CLOSED]
        resolvedBy:
          type: string
          format: uuid
        resolution:
          type: string
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...